Many changes
The Node module for connecting to Twitter was changed to Twit module.
New and deleted words from SuperVector were added to arrays in order to
update the Search vector instead of looking to the super vector in the
moment the latter is updated. This last modification was done because,
for example the Search Vector can be modify manually, so if you delete
a word from the search vector which is in the super vector, that word
then will be added each time the search vector is updated.