Limit of Search Vector phrases added
The update method of the Search Vector (SV) was updated.
1) The order of deletion and addition of words was changed. Now, words
are first deleted and then added in order to reduce the size of the SV
before adding new words

2) If the SV reach its limit of 400 phrases, the system stops adding
new words and stop the process, the words which couldn’t be added to
the SV are not deleted from the collection in order to add them after
in case some words are deleted afterwards.